Add yam command. Separate sub-parsers and arguments
